cleared backstack when navigating to MainScreen
added a popUp parameter when navigating to MainScreen so that after a user signs up and is redirected to the main screen, pressing back no longer brings them back to the signUp screen. Instead, they are exited from the app.

Thus, user has to log in to re-enter the main screen