added logic when user clicks log in button in LogInScreen
user must now have complete user details for them to be redirected to the MainScreen. When user details are incomplete, the user is redirected to the AboutDriverScreen. This prevents any problems in the MainScreen, wherein the user details are necessary.

the UserDetailsRepository is injected to the ViewModel to assist in carrying out the aforementioned logic