changed LaunchedEffect to runBlocking when getting userDetails
LaunchedEffect performs the code concurrently with the rest of the code; the code following the LaunchedEffect doesn't wait for the coroutine to complete. userDetails will still be null by the time initialState is initialized. Thus, user is always in WelcomeScreen even if they are logged in.

Instead, runBlocking blocks the current thread; the following code waits for the coroutine to complete. Thus, initial state can be properly evaluated.