displayed alert dialogs on permission denials
When the user denies the permission request once, an alert dialog is displayed indicating the importance of the permission request. If the request is launched and denied again, the user will be prompted to manually grant the permissions in the app's settings, which is launched when the user clicks the "Go to Settings" Button.

If the user does not grant permissions, Driver Mode cannot be toggled on