driver mode is toggled on when permissions are granted
Permission requests are launched when the toggle button is pressed. However, the toggle button remains off even after permissions are granted. The switch has to be pressed again so that driver mode can finally be toggled on.

In this commit, logic is added to the activity result callback. Now, granting permissions automatically turns driver mode on. We save one user press. lol.