modified alarm content to be used as entity for database table
intervalPair and nextAlarmDate are ignored because it is difficult to store non-primitive types in ROOM. Instead, those values are converted to strings, which are stored to the database

an additional constructor is added. It is called when the data retrieved from the database is converted back to an AlarmContent object.

added a main() to test the AlarmContent in a kotlin script