added random nanosecond value to alarm date-time
Random nanosecond value is added to the alarm date-time to create unique hashcode values to prevent cancelling other unrelated alarms or overwriting the database.

Alarms are set with the selected date (LocalDate), at 07:00:00.

An alarm is scheduled with a Pending Intent which requires an ID (request code) for the alarm. When cancelling, any alarms with a matching pending intent (aka ID) will be cancelled.

We decided to use the hashcode of the alarm. (Note that the alarm hashcode is also the primary key of the alarm entry in the database.)

However, when multiple alarms are set to the same day (<date> 07:00:00), they will immediately share the same hashcode and and consequently, the same ID. This would effectively cancel any other alarms with the same alarm date with the selected alarm. This would also overwrite the entry in the ROOM database, since it has the same primary key (hashcode) value.

Thus, a random nanosecond value is added to the alarm date-time (<date> 07:00:00.<randomNanosecond>) to create unique hash code values.