fix bug with crash on invalid salary or fuelCost
I initially thought I fixed the bug by putting a check in the view model event.

However, I realized that the crash occurs before the code reached the 'onSave' view model event, which takes double values for fuel cost and salary. It occurs in the screen-level composable, wherein an invalid string may be converted to a double.

To fix, I removed the properties from the event since the values passed are those that already stored in the view model.