fix bug with UI of date picker button
there was a bug wherein there would be a weird dark highlighting around the text components of the button even if the background of the button and text were the exact same color. I suspect this error is due how compose handles colors that are placed over one another. Removed the highlighting by wrapping text in surface.