set default alarm logo to app_logo_light
since app_logo is a SQUARE, it automatically applies the tint to the entire square, it doesn't know that it has an "image" inside the square.

When using a PNG drawable as an icon, it doesn't have an inherent color and appears as transparent. The tint applied by Icon() determines how the drawable is displayed. In light mode, the tint may be black, and in dark mode, it may be white.